:test_tube: Test: BoardRankingServiceTest , JournalControllerTest
