+++ server/src/main/java/com/growstory/global/utils/UriCreator.java
@@ -1,9 +1,11 @@
 package com.growstory.global.utils;
 
+import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
 
+@Component
 public class UriCreator {
     public static URI createUri(String defaultUrl, long resourceId) {
         return UriComponentsBuilder
@@ -18,6 +20,15 @@ public class UriCreator {
                 .newInstance()
                 .path(defaultUrl+"/{resource-id1}/"+anotherResourceName+"/{resource-id2}")
                 .buildAndExpand(resourceId1, resourceId2)
+                .toUri();
+    }
+
+    // 테스트용
+    public URI createUri_Test(String defaultUrl, long resourceId) {
+        return UriComponentsBuilder
+                .newInstance()
+                .path(defaultUrl+ "/{resource-id}")
+                .buildAndExpand(resourceId)
                 .toUri();
     }
 }
